backend/requirements.txt
opencv-python
dlib
face_recognition
google-cloud-storage
firebase-admin
fastapi
uvicorn
